Review issue: move test to shared/entities
